chore: Update server.hpp and server.cpp
The code changes in this commit update the server.hpp and server.cpp files. The update includes adding a new member variable `m_addr` to the `Server` class in server.hpp and implementing the `create_socket()` and `bind_socket()` functions in server.cpp. These changes are necessary for creating and binding a socket for the IRC server functionality.